requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.11
networkx>=3.2
scikit-learn>=1.4
pyarrow>=15.0
matplotlib>=3.8
PyYAML>=6.0
openpyxl>=3.1
